server/requirements.txt
flask
flask-cors
Flask-SQLAlchemy
psycopg2-binary
pytest
pytest-flask
flask_jwt_extended
python-dotenv